Wip/versioning support

+++ helm/initial_registration/initial_registration.sh
@@ -12,7 +12,7 @@ GetPathToCurrentlyExecutingScript () {
 
 		cd `dirname $TARGET_FILE`
 		TARGET_FILE=`basename $TARGET_FILE`
-		local ITERATIONS=0
+		ITERATIONS=0
 
 		# Iterate down a (possible) chain of symlinks
 		while [ -L "$TARGET_FILE" ]
@@ -39,18 +39,34 @@ GetPathToCurrentlyExecutingScript () {
 GetPathToCurrentlyExecutingScript
 set -eux
 
+VERSION=$(grep 'cordaVersion:' $DIR/../values.yaml | cut -d '"' -f 2)
+
 OUTPUT_DIR=$DIR/output
 SCRIPT=$OUTPUT_DIR/corda/templates/initial_registration.sh
-NODE_CONF=$OUTPUT_DIR/corda/templates/node.conf
+OUTPUT_NODE_CONF=$OUTPUT_DIR/corda/templates/node.conf
+NODE_CONF=$DIR/../files/conf/node-$VERSION.conf
 
-helm template $DIR -f $DIR/../values.yaml --output-dir $OUTPUT_DIR --set-file node_conf=$DIR/../files/node.conf
+if [ ! -f "$NODE_CONF" ]; then
+	set +x
+	echo "ERROR: The node.conf file could not be read, file path: $NODE_CONF"
+	echo "This most likely means you are targetting a cordaVersion in the helm/values.yaml file which does not have a corresponding node.conf file in the helm/files subfolder."
+	echo "This can in most cases easily be fixed by copying an existing node.conf file and naming it according to the version you want to use."
+	echo "Please follow this guide: If you are deploying a version 4.3, and there is only a conf file for 4.2, copy that one. If you are deplying a version for 4.2.20190221, you should also copy the 4.2 node.conf file."
+	echo "To avoid using a too new file, never copy a newer node.conf file as base, because it might contain new settings that your targeted Corda version does not know about."
+	echo "Should you run into any issues while starting up the Corda node with this node.conf file, just check the node workspace logs folder, there you will find the exact details of what the node.conf should look like in the case of errors."
+	echo ""
+	echo "Please repeat the same copy and naming steps for the bridge.conf and float.conf that exists next to the node.conf file!"
+	exit 1
+fi
+
+helm template $DIR -f $DIR/../values.yaml --output-dir $OUTPUT_DIR --set-file node_conf=$NODE_CONF
 mv $OUTPUT_DIR/corda/templates/initial_registration.sh.yml $SCRIPT
-mv $OUTPUT_DIR/corda/templates/node.conf.yml $NODE_CONF
+mv $OUTPUT_DIR/corda/templates/node.conf.yml $OUTPUT_NODE_CONF
 
 # Helm always adds a few extra lines, which we want to remove from shell scripts
 tail -n +3 "$SCRIPT" > "$SCRIPT.tmp" && mv "$SCRIPT.tmp" "$SCRIPT"
 # And from compiled node.conf
-tail -n +3 "$NODE_CONF" > "$NODE_CONF.tmp" && mv "$NODE_CONF.tmp" "$NODE_CONF"
+tail -n +3 "$OUTPUT_NODE_CONF" > "$OUTPUT_NODE_CONF.tmp" && mv "$OUTPUT_NODE_CONF.tmp" "$OUTPUT_NODE_CONF"
 
 # Make the script executable
 chmod +x $SCRIPT
